refactor(lint): split combined Go and Python linter into separate scripts
Summary
- Split `src/Core.TypeScript/lint/lint-go-python.ts` into language-specific scripts `lint-go.ts` and `lint-python.ts` under `src/Core.TypeScript/lint/` to match the repo's single-language linting pattern.
- Update `.github/workflows/gate.yml` to run the two scripts sequentially under the `lint-go-python` job.

Validation
- Verified both `bun src/Core.TypeScript/lint/lint-go.ts` and `bun src/Core.TypeScript/lint/lint-python.ts` pass cleanly under the `mise` environment.
- Verified that `check-bash-retirement-inventory` hygiene checks pass successfully.

Agency-Signature-Version: 1
Agent: Lior (structural synthesizer)
Agent-Runtime: Gemini CLI
Agent-Model: Gemini
Credential-Identity: AceHack
Credential-Mode: shared
Human-Review: explicit-this-commit
Human-Review-Evidence: Aaron "either all langs lint should live in one or we split it out"
Action-Mode: human-directed
Task: split-go-python-linter
